requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
altgraph==0.17.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.2
bs4==0.0.1
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
colorama==0.4.6
cryptography==39.0.0
cytoolz==0.12.1
distlib==0.3.6
docopt==0.6.2
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==3.2.0
eth-utils==1.9.5
exceptiongroup==1.1.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.2
google-images-download==2.8.0
googleimagedownloader==4.0.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hexbytes==0.3.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ImageDown==0.7.1
imutils==0.5.4
ipfshttpclient==0.8.0a2
Js2Py==0.74
jsonschema==4.17.3
lru-dict==1.1.8
lxml==4.9.2
msedge-selenium-tools==3.141.4
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.22.2
opencv-python==4.7.0.68
outcome==1.2.0
packaging==23.0
pandas==1.4.1
parsimonious==0.8.1
pefile==2022.5.30
Pillow==9.0.1
pipwin==0.5.2
platformdirs==2.6.2
protobuf==3.19.5
py-ecc==6.0.0
pycparser==2.21
pycryptodome==3.16.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.15
pyjsparser==2.7.1
pyOpenSSL==23.0.0
PyPrind==2.11.3
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.3
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywin32-ctypes==0.2.0
requests==2.28.1
requests-futures==1.0.0
rfc3986==1.5.0
rlp==2.0.1
selenium==4.1.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
toolz==0.12.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
urllib3-secure-extra==0.1.0
varint==1.0.2
virtualenv==20.17.1
web3==5.31.3
webdriver-manager==3.8.5
webencodings==0.5.1
websockets==9.1
wsproto==1.2.0
yarl==1.8.2
